// app/src/api/jInshanBodyShopApi.schemas.ts
export interface Product {
  productId?: string;
  /** @nullable */
  name?: string | null;
  price?: number;
  /** @nullable */
  description?: string | null;
  weight?: number;
  category: ProductType;
  quantity?: number;
  withdrawn?: boolean;
  shop?: ProductShop;
}

export interface ProductAndQuantity {
  productId?: string;
  quantity?: number;
}

export interface ProductApiGetReturnType {
  maxPages?: number;
  /** @nullable */
  items: Product[] | null;
}

export type ProductShop = (typeof ProductShop)[keyof typeof ProductShop];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ProductShop = {
  NUMBER_0: 0,
  NUMBER_1: 1,
  NUMBER_2: 2,
} as const;

export type ProductType = (typeof ProductType)[keyof typeof ProductType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ProductType = {
  NUMBER_0: 0,
  NUMBER_1: 1,
  NUMBER_2: 2,
  NUMBER_3: 3,
  NUMBER_4: 4,
  NUMBER_5: 5,
  NUMBER_6: 6,
  NUMBER_7: 7,
} as const;